Fix build with -Wvarargs
va_args is picky about the type of the function argument which gets
passed, and bool gets promoted to int by default, leading to undefined
behavior.

I took the liberty to rename 'direct' to 'is_direct' to keep the
original intention now that the type is int -- feel free to change this
if you disagree.

Here's clang's error message:

  ../tests/md_test.c:377:22: warning: passing an object that undergoes default argument promotion to 'va_start' has undefined behavior [-Wvarargs]
        va_start(va, direct);
                     ^
  ../tests/md_test.c:362:38: note: parameter of type 'bool' is declared here
  md_test_dep(md_dep_type_t type, bool direct, ...)