C++: Forbid instantiation of Logs
Now that the relevant methods are logging, there's no point in letting
the user code instantiate this (or in instantiating this at all).

I had to re-run cmake locally in order for this change to be picked up
by SWIG. Without that step, I was getting compiler errors like this one:

  swig/lua51/libsysrepoLua51LUA_wrap.cxx:4755:26: error: call to deleted constructor of 'Logs'
      result = (Logs *)new Logs();
                           ^
  ../swig/cpp/src/Sysrepo.h:106:5: note: 'Logs' has been explicitly marked deleted here
      Logs() = delete;
      ^

I suspect that there is some missing dependency somewhere in the CMake
setup for SWIG generation.

I tested this to make sure that the Python2 bindings still work:

--- a/swig/python2/examples/pytho2_application_example.py
+++ b/swig/python2/examples/pytho2_application_example.py
@@ -99,6 +99,7 @@ try:
     else:
         print "\nYou can pass the module name to be subscribed as the first argument"

+    sr.Logs.set_stderr(4)
     # connect to sysrepo
     conn = sr.Connection("example_application")