Logging: Do not skip diff computation when a log callback is registered
This could be a bit controversial -- the current code takes a shortcut
whenever the debug logging is not active on syslog or stderr sinks.
However, because SR's code has no way to know about the log level
processed by the logging callback, it is safer to perform this
maybe-useless computation and to pass it to the callback. Otherwise
there's potential for information loss, which is not pretty, IMHO.

An alternative would be to employ log level filtering for user's
callbacks, too.