Install the C++ headers
I'm not installing directly to include/sysrepo because the include setup
is a bit weird. There's an include/sysrepo/*.h and include/sysrepo.h,
and the include paths not necessarily reflect all that.

Long term, I would like the C++ headers to use better include paths,
such as sysrepo/Foo.h, but that would require some shuffling around and
SWIG updates, and it's too late in the evening for them now.