cmake: use GNUInstallDirs for standard install paths
Warning: on non-Debian amd64 systems, this will change the installation
directory of libraries, they will be put into
${CMAKE_INSTALL_PREFIX}/lib64/ now. The old versions of libraries which
might still be in lib/ will be left in place, which means that there's
some potential for breakage on developers' systems. Still, I think that
following cmake conventions is a good reason to follow suit here.