cmake: RelWithDebugInfo is not a debug build
The RelWithDebugInfo mode is intended to produce a release binary with
all optimizations, but with the debug info still present (maybe to be
moved to a -dbg package by Linux distros). This means that these build
configurations should use global, systemwide /etc/sysrepo/ for repo
storage.

A simple regex match for "debug" matches RelWithDebugInfo, too.

Testing:
- no extra options -> debug due to manual override earlier in
  CMakeLists.txt
- RelWithDebugInfo -> release
- Debug -> debug