C++: Use an extra class for all sysrepo exceptions
There is value in knowing that a given exception originated within the
sysrepo -- especially in users' exception handlers which might get
std::runtime_error instances from both SR and some unrelated libraries.
And because of the way how the RTTI works in C++, the destructor has to
be out-of-line in a separate translation unit.

With the captured error_code, one could override std::exception::what
and construct the string in that context. However, I chose not to do so
because a std::runtime_error has no default constructor anyway, and I
would have to come up with some string just to throw it away later.