cpp: rebase code to use C++'s friend instead of public pointer methods
With this refactoring the helper methods for getting C pointer of
structs in C++ is replaced with direct access to under laying C struct
with the friend C++ feature. Not all instances are replaced. Only in
Session class remains get() because it is used in python/lua callbacks
and the refactoring is a bit more complex.

Signed-off-by: Mislav Novakovic <mislav.novakovic@sartura.hr>