tests: make sure that deamon was really killed
When killing a sysrepo/sysrepo-plugin deamon, we should wait for removing
its PID file. A consequent test can check for the PID file and expect that
the daemon is runnig while it is shutting down and the PID file is going
to be removed in a while. Such an expectation can lead to a failure of the
consequent test.