tests: workaround for glibc bug in fork
https://bugzilla.redhat.com/show_bug.cgi?id=1275384

Previous solution was not good enough for the case the test is running
under another process which watch it (valgrind or ctest). It is
necessary to catch the SIGSEGV signal, hide it and stop the process.
The code from the previous solution to this bug then tries to run fork
again.