Make it possible not to run sysrepoctl during `make install`
When cross compiling, one cannot easily run the produced binaries (e.g.,
when compiling ARM target binaries on an x86_64 host node). This change
makes it possible to produce a shellscript which can be manually copied
to the target (or integrated into the image-making process, or
whatever). When executed, it ensures that all YANG files are installed
as needed.

Until this script has been run, sysrepo's installation is unusable. You
cannot run sysrepod, in particular.

fixes #738, fixes #748, fixes #757