fix traversing schema tree when exploring modules dependencies
Ignore the augment nodes and go directly through them because the same
way we got into (via the child pointer), we have to get back (usually
parent, but in case of augmenting nodes we have to go through the agment
node). This applies when augment targets the same module (or the submodules
from the same module) because only in this case we get into the augment.
In case of augments targeting different module, the augment are processed
separately and it is necessary to detect this situation and stop at augment
node when going back through the tree.

Fixes sysrepo/sysrepo#671