Github templates for issues and pull requests.
This is a suggestion for github templates and issues.

To better use of developers time, a filter is needed for repetative and/or basic issues.
It's no fun telling users same thing over and over (like attach log, steps to reproduce, did you check devel, etc.).

We're thinking that this current version has too much text, but it also seems necessary.
And it could still grow. Some kind of readme file would be better place for much of it. Headers would be left for issuel to fill out.

Signed-off-by: Antonio Paunovic <antonio.paunovic@sartura.hr>