sysrepocfg: Use SR's helper for setting XPath values
The old method was not freeing the memory stored into the XPath and the
actual value data. At first, I tried to fix that by the usual
`sr_new_val`/`sr_free_val` combo. This requires heap allocation for the
whole `sr_val_t`, but it's aligned with what the rest of the code is
doing.

However, I believe that there's no point in duplicating the functionality of
`sr_set_item_str`, so I'm making use of it as well. The logic is changed
a bit to ensure that the externally observed behavior (and the return
code) remains unchanged. A special case for changing lists is preserved.