sysrepocfg: Do not validate the existing datastore content
When we're trying to modify the existing datastore by tweaking some
nodes in the tree, the state of the existing DS is arguably not
important. If it is not correct as-is, there is no chance of making the
matters worse by any edit because it's broken already. We *might*
actually improve the situation by an edit which makes it valid again.

This improves the overall performance of `sysrepocfg --set ...
--set-value ...` as measured by wall time (and by CPU cycles as measured
by `perf`) by about 17% in my testing.  I'm using an in-process
`sysrepod` and all of libredblack/libyang/sysrepo have been built with
ASAN). I've also checked with `perf` that the actual time spent in all
invocations of `lyd_validate` has been reduced (again based on
`cycles:u`).

cc #892