Support handling of pending events with a custom event loop
When sysrepo is running with its own event loop, it ensures that all
pending events have been processed when the subscription manager gets
destroyed. This patch adds an equivalent functionality for scenarios
where the application provides its own event loop.

This functionality requires a real blocking call, so it is not enough to
reuse the existing event handling that is based on asynchronous writes
to a notification pipe. A new callback function was needed.

I had to add both a new global static to `client_library.c` and an extra
member to the `cl_sm_ctx_t` because `sr_fd_watcher_init` and
`cl_sm_cleanup` are defined in separate translation units.

Technically speaking, it's still possible to call `sr_fd_watcher_init`
with a NULL synchronizing callback. However, doing so results in some
events not being handled if, e.g., the event loop runs in a different
thread and the subscription manager gets destroyed "too soon", prior to
the event loop having handled "everything".