cross-building: Workaround calls to sysrepoctl/sysrepocfg
Sysrepo has always wanted to call binaries which it has just produced.
That works well when building natively, but it's a no-go when doing
cross-compilation.

I've added a workaround earlier in commit 00025c49. One could use two
options, CALL_TARGET_BINS_DIRECTLY and INDIRECT_YANG_INSTALL_DIR to
produce a shell script which could be used to initialize /etc/sysrepo
later on, perhaps on the target system. However, that is still not a
complete fix for scenarios which intend to, e.g., produce a
self-contained disk image to be flashed to an embedded device.

I'm using Buildroot which makes it super-easy to build a package twice
-- once as a native, host package, and the second time as a cross-built
package targeting another system with a possibly different architecture.
This patch uses that feature. One can first build a special "throw-away"
sysrepo which will only be used for its `sysrepoctl` and `sysrepocfg`.
This host-side-only sysrepo should be built with an appropriate
REPOSITORY_LOC to point to the target's /etc/sysrepo.

When building the second sysrepo that is supposed to run on the target,
proceed as normally, but pass two additional arguments,
CALL_SYSREPOCTL_BIN and CALL_SYSREPOCFG_BIN which point to the throwaway
sysrepo. That way, these tools can be used for executing host's native
code which however installs modules into the target's /etc/sysrepo.
These tools can be also used for other packages such as Netopeer2.