plugind: Fail if sysrepod is not available
I have to admit that I am no fan of automagic dependencies; I believe
that it is very easy to configure one's init system with proper service
dependencies (yeah, see the previous commit to check how great I am at
that</sarcasm>). I don't see a reason why an attempt to start
`sysrepl-plugind` should succeed in case `sysrepod` is not running, or
has not been configured to run.

Based on a review from Michal Vaško, I added an option which activates
the old behavior. Use the `-D` flag to ensure that `sysrepod` is
auto-started if needed.