Add idrefs as ext deps even when the src mod is not implemented yet
The following model graph is used to exhibit the problem:
```
      +-------------+
      |             |
      | id-ref-base |          +------------------+
      |             |          |                  |
      +----+-----+--+          | id-def-extension |
           ^     ^             |                  |
           .     | Augment     +------+--------+--+
           .     |                    |        .
           .     |                    |        .
    Extend .     |                    |        .
           .     |            Import  |        . Extend
           .     |                    |        .
           .     |                    v        v
       +---+-----+--+   Import      +-+--------+--+
       |            +-------------->+             |
       | id-ref-aug |               | id-def-base |
       |            +<..............+             |
       +------------+    Extend     +-------------+
```

If the model installation order is:

 1. `id-ref-base`
 2. `id-ref-aug`
 3. `id-def-base`
 4. `id-def-extension`

then a dependency will be left out of the graph and the user will be
unable to configure the identity defined in `id-def-extension`.

Before this change, a consumer of an identity (i.e. a module with an
identityref) would only add the source of that identity module as an
extension to itself when that source is marked as 'implemented'. However
this causes that dependency to be lost, as it won't be regenerated when
the source of the identity is later installed with sysrepoctl.

An alternative implementation to this could be to regenerate the
identityref module when the identity source module is installed. However
this seems like it would introduce quite a bit more complexity into the
installation process so we've opted for this approach in this patch.

A new unit test in rp_dt_edit_test was added to cover this use case.

A minor change was made to an existing test in md_test to reflect the
new behavior.