Fix segfault due to NULL child (aug w/ feature)
Installing a YANG module with sysrepoctl results in a segmentation
fault if the module contains an augment whose children are contingent
upon a feature. The same segmentation fault occurs if the module
contains an augment with a uses clause where the corresponding
grouping’s children are contingent on a feature. Issue #1105
documents this issue. The included test case demonstrates the failure.

To address this, `md_traverse_schema_tree()` has been modified to
tolerate the return of a null child from the call to the function
`lys_getnext()`.