[RFC] access-control: disable identity change when nacm is enabled
DISCLAIMER: THIS IS NOT A PROPER FIX BUT MORE OF A WORKAROUND. I propose
this patch as a starting point to base the discussion on.

Sysrepo has two different mechanisms to manage what user may access
which data:

- "access control": based on UNIX file system permissions. The user
  which is identified by sysrepo must have read (and/or write)
  permissions to the files in /etc/sysrepo/data. This is suitable for
  local access via the sysrepocfg tool but not quite for remote NETCONF
  access (where there might not even be a corresponding UNIX user).
- "nacm": based on RFC 8341, it is implemented by sysrepo and uses the
  ietf-netconf-acm YANG module as configuration data. It allows more
  fine grained control over what parts of the configuration a user has
  access to (read/write/execute).

NACM is disabled by default and must be enabled at compile time.
Surprisingly, enabling NACM does not disable the file system permissions
based "access control".

When relying on NACM for access control and sysrepo is a backend for
Netopeer2, if the NETCONF user does not have UNIX permissions to modify
the files under /etc/sysrepo/data, he/she cannot modify parts of the
configuration (even though the NACM rules says he/she can).

Giving filesystem write access to every user that need to modify part of
the configuration is redundant with adding access rules in
ietf-netconf-acm. Also, it allows a coarser grained access control and
it may not be desirable for security reasons.

Disable identity switch for file system checks so that they always
succeed when NACM is enabled at compile time.

A proper way to do it would be to disable the identity switch only when
the user is connected via NETCONF and that NACM is enabled (and
enforced). However, this involves changing the API of multiple functions
in access_control.h and I don't know the code well enough to do that
without breaking everything.