persistence_manager: only write <enable-running /> for supported subscriptions
When accessing the sysrepo protobuf API without using the client
library, it is possible to send a RPC_SUBS message with
enable_running=True. Sysrepo accepts the subscription and adds it in the
persistant-data with a <enable-running /> tag.

This leads to an error when sysrepo tries to read the file again and
parse it with libyang:

[ERR] libyang error: When condition "../type = 'module-change' or ../type = 'subtree-change' or ../type = 'dp-get-items'" not satisfied.
[ERR] Parsing persist data from file '/etc/sysrepo/data/example.persist' failed: When condition "../type = 'module-change' or ../type = 'subtree-change' or ../type = 'dp-get-items'" not satisfied.

Make sure to respect the condition in the yang module.

Signed-off-by: Robin Jarry <robin.jarry@6wind.com>