C++:  do not silently ignore some exceptions
We were using a static list of error codes with no fallback for
unknown values and no warning for unhandled, newly added error codes.

We don't need a whitelist, so let's simplify the code to always
pass the error code to the exception constructor. The exception
constructor can handle unknown errors.

Signed-off-by: Mislav Novakovic <mislav.novakovic@sartura.hr>