request_processor: ensure that request id is growing
We want to use the request id to identify a GET_ITEMS request (which is
split in multiple requests). At the moment, it does not work, because the
request id is growing only in a session. If two sessions are made for one
request, we have no way to know that the second request is a new one,
because they have the same id.

This patch ensures that we only have one source for request id for all
sessions. As it was done before, we use the number of request sent, but
this time it is a global counter, not a per-session counter.

Signed-off-by: Samuel Gauthier <samuel.gauthier@6wind.com>