add request id to sr_dp_get_items_cb
When a data request is made, sysrepo will ask to a subscriber for each level
under the subscription, calling the subscriber's callback once per layer.

There are case when this is a problem, because we can not be sure that the
data is consistent, as it was asked in several separate requests.

This commit adds request_id to the sr_dp_get_items_cb prototype. Then, the
subscriber can know if his callback is called for one state data request
or several by checking the given request id. The subscriber can take the
necessary mesure to ensure the data is consistent.

Signed-off-by: Samuel Gauthier <samuel.gauthier@6wind.com>