Support notifications tied to data nodes
These changes fix a few cases where sysrepo was not processing notifications
connected to a data node. (Support for this was added beginning in RFC 7950.)

Summary of changes:

* `LYS_NOTIF` cases were added to `sr_copy_node_to_tree_internal()` and
`rp_dt_get_value_from_node()`. Previously, these functions were treating this
node type as unsupported and returning with an error.

* `rp_dt_get_values_from_nodes()` was modified so that it does not produce a
`values` array with blank entries for skipped nodes (like notifications).
Prior to this fix, it was correctly reporting the number of non-blank values
in the array via `value_cnt`, but the array itself would actually be larger
due to the blank (skipped) entries.

We've written integration tests that demonstrate these types of notifications.
This commit, together with a forthcoming commit to libyang, will allow those
failing test cases to pass. (See [CI results], lines 7282-7284, or
[test source code].) These test cases cover several different variations of
notifications tied to data nodes.

[CI results]: https://travis-ci.org/ADTRAN/netopeer2-integration-tests/builds/414144146#L7282
[test source code]: https://github.com/ADTRAN/netopeer2-integration-tests/blob/master/tests/test_notif.py#L229