Validate inverse data dependencies on commit
Fixes #1242

Whenever a module is changed, any other modules that have data
dependencies on that module should also be checked to ensure their
data references are still valid.

There is a bit of a performance penalty for doing this, and ideally
one would only check the inverse deps if the specific nodes being
referenced were changed. That is an optimization that could be done in
the future.