add original xpath to sr_dp_get_items_cb
When a data request is made, sysrepo will ask to a subscriber for each
level under the subscription, calling the subscriber's callback once per
layer.

When there is a lot of layers, it can become a problem, because a lot of
requests are made between the subscriber and sysrepo.

This commit adds original_xpath to the sr_dp_get_items_cb prototype. The
subscriber can then reply only for the nodes that are needed.

Note that the 'default' case of np_data_provider_request only happens during
unit tests.

Signed-off-by: Samuel Gauthier <samuel.gauthier@6wind.com>