lua: remove bindings
we are deprecating the Sysrepo Lua bindings.
With the current changes to the Sysrepo C++ bindings we had to use
simple wrapper class (ptr class) for the objects returned and sent to
the Lua bindings which makes them a bit awkward to use now, the
example below:

-            change = sess:get_change_next(it)
+            change = sr.ptr(sess:get_change_next(it))

We used Lua bindings for scripting on OpenWrt alongside with Python
but now Python is better suited for that. Besides us, we aren't aware
of any one using the Lua bindings and they are making the C++/Python
bindings harder to maintain.

Signed-off-by: Mislav Novakovic <mislav.novakovic@sartura.hr>