tests: Use CMake's native timeout
The test suite used two non-standard mechanisms for setting a timeout
for each test:

- if the `timeout` binary was available, all tests were wrapped within
that and with a timeout of six minutes,
- a custom watchdog thread was running, killing the process after a
configurable timeout which was five minutes for all tests except
nacm_cl_test where it was a bit more than eight hours

Given that CMake has a native feature for this, let's use that one and
remove both custom methods.

The timeout for `nacm_cl_test` was set to 30'000 seconds (that's about a
working day). That looks like a nonsensical value, so let's set it to 10
minutes and hope for the best.