Ensure that feature test macros work correctly
These feature test macros are required to be defined prior to including
any standard headers. This is a transitive requirement because headers
are, in general, free to include other headers. The sysrepo's code
violated this because it included <stdatomic.h>.

This fixes a build failure when using clang on Fedora 29. Apparently,
the order of #include is different from using GCC. By the time
sr_utils.c defined _GNU_SOURCE, the <features.h> was already processed,
and therefore _GNU_SOURCE was not turned into __USE_GNU (which is what
<bits/fcntl-linux.h> uses as a guard for the F_OFD_GETLK etc.