Don't send notifications back to the session that changes database
If a single threaded process is subscribed to validate the database and
then commits to the database it will deadlock/timeout as the validate
callback can never be run. In this case it could be assumed that the
process has already validated this data before committing it to the
database so it would be useful if it was possible for this functionality
to work without deadlocking. To achieve this we now store the session id
that subscribed to validate the database and if this session id matches
the id of the session that has committed to the database then we skip
the subscription notification - avoiding the deadlock.