common BUGFIX support for decimal64 in union
... when converting from sr_val_t to libyang.
However, the support is only basic and there are
use-cases when it will not be resolve correctly
and much more cases when the value will simply
change. That is the result of trying to make the
conversion behave as expected for most common
use-cases. It is not possible to implement perfectly
because of double precision loss.

Refs #1880