sub shm BUGFIX SHM re-lock checking
Whenever subscriber locked the SHM,
started processing an event, unlocked
it, and then locked again, it must be
checked that no unexpected changes occured
such as another subscriber writing an
error or simply a timeout from the originator.
Also, write lock is needed when remapping
the SHM to a specific size as it can be shrinked
effectively removing some data from its end.

Fixes #1949