tests BUGFIX increase timeout in test_process
Under TSan, this test sometimes hangs on wait(), because the forked
process never finishes and spins on some sort an init and/or cleanup
procedure, that I'm not able to debug. This is reproducible consistently
on our CI (a heavily multi-threaded environment). On my work laptop, I
was able to reproduce this by setting CPU clock speed to a low value.
TSan creates a huge overhead (and slows the forking process), which
seems to be mandatory for this issue to happen. Lowering the timeout to
a very low value (like 1) also increases the chance this appears. This
makes me think that the issue appears, if the timeout expires at some
critical moment, for example, when the forking is not yet done.

This is definitely not a proper fix as I'm not sure what exactly is the
reason this happens (it could be a deeper problem), but increasing the
timeout solves it for our CI (or at least most of the time, I haven't
gotten the issue since I increased the timeout). Also, I think the
nature of the test

The test now runs for around two seconds longer because of this, but
that shouldn't be a problem since the patch that enables the tests to be
run in parallel.