c++ BUGFIX oper_data output
If subscribing to a list type top-level node, all siblings need to be
duplicated.

Also, I deleted a line which was (probably) supposed to create a nullptr
shared_ptr, but instead created a valid libyang::Data_Node with an
invalid lyd_node inside. That broke the duping algorithm if the `parent`
variable wasn't reassigned inside the callback.