Do not let SR_SUBSCR_DONE_ONLY mask out SR_EV_ENABLED events
The SR_SUBSCR_DONE_ONLY event is documented to control whether
SR_EV_CHANGE/SR_EV_ABORT are sent. Since the rewrite, it used to also
skip generating the initial SR_EV_ENABLED as well. That's surely a
mistake because that one is supposed to be gated by the SR_EV_ENABLED
flag.

Fixes: f36f312e ("sysrepo CHANGE add special enabled event")