Work around a paranoid kernel protection for sticky-bit directories
Since the Linux kernel 4.19+ and systemd v241, there's a new sysctl knob
which prevents users from overwriting "someone else's files" that are
located in world-writable directories with the sticky bit set. That goes
directly against sysrepo's permission model (which might become
irrelevant anyway because it sits in parallel to the NACM).

When this change was made upstream, they cited a list of CVEs which are
supposed to be preventable via that change. I believe that it turns out
that all of them are actually *other*, serious bugs, but I acknowledge
that it's true that `fs.protected_regular = 1` can indeed help against
something stupid like `exec /tmp/install.sh`:

- CVE-2000-1134: Insecure file creation under /tmp.
- CVE-2007-3852: Insecure file creation of /tmp/sysstat.run.
- CVE-2008-0525: Symlinks below /tmp.
- CVE-2009-0416: Symlinks below /var/tmp.
- CVE-2011-4834: Symlink attack, and then content spoofing via a FIFO.
- CVE-2015-1838: Download & execute /tmp/install.sh.
- CVE-2015-7442: Insecure file creation in /tmp.
- CVE-2016-7489: Insecure file creation in /tmp.

Looking at that list, I still think that the upstream decision is
nonsense, but hey, I've been postponing that mail to linux-fsdevel for
four months already, so let's solve this within sysrepo. Just provide a
drop-in file for systemd to undo this nonsensical (IMHO, anyway) policy
change.

Bug: https://github.com/sysrepo/sysrepo/issues/2080