Add a helper for NETCONF- and RESTCONF-compliant errors
Sysrepo uses a flexible internal mechanism which client applications can
use for storing arbitrary error information. So far, however, the only
conventions for how these error messages are supposed to be created were
defined in Netopeer2's README file. This patch builds on top of that
format.

I changed serialization of the error-data construct. Previously, it
would be a sequence of pairs of strings which was used for building
anyxml elements and their string values, respectively. The proposed
format accepts a libyang data tree and uses the LYB formatting for
passing that around.