Fix build when not using CMake's make generator
CMake is a generator which can be used with other tools than just
`make` -- `ninja` is a good candidate which tends to build a little
faster. It's therefore a bug to try invoking `make` on some target
directly.

Since the test fixtures only work on top of other tests, not targets,
the `tests_clean` needs to become a test. In CMake, the `add_test()`
supports just a single `COMMAND` statement (this is different from the
`add_custom_target`). Prevent duplication of code by storing the actual
command into a variable.

Fixes: 08659bdf tests FEATURE clean all files created by tests