Do not trigger pull callbacks with SR_SUBSCR_ENABLED
Currently if module_change_subscribe() is used to subscribe to a module
in the operational datastore, all operational data pull callbacks are
invoked for the module (even if not included by the subscription XPath)
if SR_SUBSCR_ENABLED is used.

This is not very intuitive since you cannot subscribe to changes to
pull data, only push data, and the change callbacks are unlikely to
be expecting the pull data. So this causes un-necessary system load
to collect data that is not used.

Therefore we prevent any pull callbacks from being invoked by using
SR_OPER_NO_SUBS when collecting data for the enabled callback.

Fixes #2651