NACM: Use sr_error_info_t for returns
The callbacks do not use this, since they need to return int. Some
functions aren't able to return errors, so I didn't change the return
value for those.

Since the sr_nacm_strarr_sort_add is used in the callbacks, it also
doesn't return an int. I have removed the EMEM macro from there,
hopefully, the return value is enough. Another option would be to add
the sesssion argument into the function.

sr_nacm_getpwnam: This function returned three values, 0 for success, 1
for "user not found" and -1 for error. The new sr_nacm_getpwnam uses
sr_error_info_t as a return value and uses the SR_ERR_NOTFOUND code to
for "user not found". This means that the caller has to free the error
afterwards, hopefully that's not a big problem (and hopefully it's
correct).

sr_nacm_check_yang_push_update_notif now gains a return value and
propagates errors.